give 'check' events an event type
Since `put` can both create and delete configs, it seems better to keep
them all using the same event structure, even if 'check' can only emit
one type of event (`discovered`).

The alternative to this would be to split deleting out of 'put'. At that
point they would each emit a single type of event, and the interface
would be less open-ended. There's value in that, but that value has to
be weighed against the value of being able to create + delete in one
action. It's also not super clear that all resource types would support
'delete'. (But then, not all of them will support 'put', either.)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Nick Hindley <nicholas.hindley@yoti.com>