add 'delete' action, unify action interface
this removes 'event types' from the response from each action, and
now they all have the same response.

event types were needed because 'put' could both create and delete, and
we needed a way to tell the difference. so instead, we've split 'delete'
out, and now each action has only one way to interpret the events, so
the types are no longer needed.

this makes the interface a bit less flexible, but it seems worth seeing
how far we can go with this constraint for the sake of having a simpler,
more focused interface. if we can go the distance with this limitation
it feels like stronger validation of the resources concept.

along the way I've put the configuration for each command under an
'actions' field in the info response. i'd like to also have it so
actions are optional, i.e. you can have a resource that doesn't
implement 'delete' or 'put'. this seems like it'll be necessary if we
want a common resource interface to be used for things like
notifications or triggers where there's nothing to really delete, and
even today there are cases where resources are 'read-only' and don't
support put at all (e.g. RSS).

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>