execute default command for image, not ./info
we were already relying on image semantics here by expecting it to be in
the working directory, so relying on CMD should be no worse.

this will allow greater flexibility for resource type authors, allowing
for a single-binary entrypoint rather than forcing them to have an
'info' executable in a particular location.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>