bring TLS configuration into the interface
this will allow us to have a common mechanism for configuring TLS across
all resource types. having it be a formal part of the interface feels
sensible as all resources are meant to interact with external state -
and in doing so, TLS will be a commonplace.

a later RFC (or perhaps another proposal within this one) will
demonstrate how users configure TLS - it may be something like the
pipeline-level, or another field alongside 'source' on the resource
definitions.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>