remove proposals for resource interpretations
each of these proposals should be submitted as separate RFCs. each RFC
should show how the v2 *and* v1 interfaces are to be interpreted.

i've also removed the artifact resources proposal, which I initially
included in order to provide a concrete example relating today's
resource usage to the v2 interface. however, it can technically be done
independently of the v2 interface, if not just as a way of achieving
'explicit `get` after `put`' in a backwards-compatible way. so let's
just propose it separately and link to all four to demonstrate how it is
just one of four proposed interpretations.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Nick Hindley <nicholas.hindley@yoti.com>