allow '.', identifiers must start with a letter
The decision to allow '.' feels somewhat arbitrary but I feel is easily
justifiable from the standpoint that it is often used for CI/CD related
things like version numbers or a DNS name identifying a deploy target.
It is also one of the few allowable special characters that does not
require any additional URL encoding.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Nick Hindley <nicholas.hindley@yoti.com>