revise name collision and unarchiving semantics
A lot of things become easier by just having everything exist in the
same namespace.

All of the same navigation constructs continue to work, and there is no
concern of a name collision when un-archiving.

If someone wants to re-use a name, it might be better to just force them
to give a clearer name to the archived pipeline so that it can be
distinguished from the new one.

This also introduces un-archiving by way of setting the pipeline again.
This way a `set_pipeline` step can be commented-out (or an instance can
go away) and then the pipeline will just come back as it was when the
pipeline becomes set again.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Nick Hindley <nicholas.hindley@yoti.com>