move RFC assignment to 'submitting an RFC' section
the 'reviewing an RFC' section will be linked to by the Concourse
website, and is aimed at prospective RFC reviewers.

this RFC assignment part is directed at RFC authors as part of
explaining the process, so it should exist in the 'submitting an RFC'
section instead.

tweak wording and fix some typoes along the way.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>